pyproject.toml
[tool.poetry]
name = "repl_python3_Bott"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
discord = "^1.7.3"
google-api-python-client = "^2.31.0"
google-auth-oauthlib = "^0.4.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
